# feat(groupComparison): Enable parallel processing for groupComparison function