--- a/R/groupComparison.R
+++ b/R/groupComparison.R
@@ -7,7 +7,7 @@
 #' @param log_base base of the logarithm used in dataProcess.
 #' @param numberOfCores Number of cores for parallel processing. When > 1, 
 #' a logfile named `MSstats_groupComparison_log_progress.log` is created to 
-#' track progress.
+#' track progress. Only works for Linux & Mac OS.
 #' @inheritParams .documentFunction
 #'
 #' @details
@@ -114,7 +114,7 @@ MSstatsPrepareForGroupComparison = function(summarization_output) {
 #' @param samples_info data.table, output of getSamplesInfo function
 #' @param numberOfCores Number of cores for parallel processing. When > 1, 
 #' a logfile named `MSstats_groupComparison_log_progress.log` is created to 
-#' track progress.
+#' track progress. Only works for Linux & Mac OS.
 #' 
 #' @importFrom utils txtProgressBar setTxtProgressBar
 #' @importFrom parallel makeCluster clusterExport parLapply stopCluster