--- a/man/dot-groupComparisonWithMultipleCores.Rd
+++ b/man/dot-groupComparisonWithMultipleCores.Rd
@@ -26,7 +26,7 @@
 
 \item{numberOfCores}{Number of cores for parallel processing. 
 A logfile named `MSstats_groupComparison_log_progress.log` is created to 
-track progress.}
+track progress. Only works for Linux & Mac OS.}
 }
 \description{
 Perform group comparison per protein in parallel