… NAMESPACE
… man/dot-groupComparisonWithSingleCore.Rd