# Motivation and Context

In theory, MSstats should produce more meaningful results if clients set up experiments with more bioreplicates.  However, as clients begin to input larger datasets into MSstats, they have been running into issues regarding speed & memory.  This pull request aims to speed up the MSstats groupComparison function by adding a parameter called `numberOfCores` that should allow users to process protein comparisons in parallel.  

## Changes
- Add new dependency on `parallel` package and modified documentation accordingly.
- Added new `numberOfCores` parameter to groupComparison function.  Default is 1 core, which goes through the regular `for` loop approach with a progress bar indicator.
- Used the parallel package's parLapply function to process `MSstatsGroupComparisonSingleProtein` in parallel per protein.  
- Note that for parLapply, it does NOT allow for a progress bar indicator.  
    - There is a package called `pbapply` that sets up a progress bar indicator with `parLapply`, but it runs much slower than the `for` loop implementation since different cores need to constantly communicate with each other to update the progress bar. Still evaluating if I can finagle something with `pbapply` to speed up the computation despite the progress bar.
    - To handle progress, I set up a logfile called `MSstats_groupComparison_log_progress.log` that each core can write to every single time it encounters a protein index divisible by 100 - it's not perfect, but as input dataset gets larger, the log messages give a good approximation on how many proteins have been processed.

## Testing
- Tested speed on this [2.3 GB dataset](https://drive.google.com/drive/u/0/folders/1DwRQcyrk9CylytMEv3X_YyPAsWTnI84L) and encountered a ~76% increase in performance from 1 core to 4 cores on my M1 mac that has 8 cores (i.e. 30 seconds -> 17 seconds).  Have not gotten a chance to test on Windows/Linux yet.
- Added unit tests for groupComparison verifying that parallel processed function emitted the same number of rows as groupComparison with default parameters.  
- Manually verified that if you set the numberOfCores parameter to greater than the number of cores on your device (i.e. by calling detectCores()), no issues occur.  


## Checklist Before Requesting a Review
- [x] I have read the MSstats [contributing guidelines](https://github.com/Vitek-Lab/MSstats/blob/master/.github/CONTRIBUTING.md)
- [x] My changes generate no new warnings
- [x] Any dependent changes have been merged and published in downstream modules