@tonywu1999 2024-02-15
As discussed offline, here's next steps for the next revision:

- Improve error handling: how do we handle one core having errors and its impact on the other cores?
- Warning handling: mclapply throws errors upon warnings.  We need to make sure parallel backend doesn't treat warnings as errors
- Operating systems: need to test on Windows and Linux.
- Filesize: run tests on performance on with and without file progress saves.