@tonywu1999 2024-02-16
### Progress File Writing Test

- Ran tests with 4 cores on this [2.3GB dataset](https://drive.google.com/drive/u/0/folders/1DwRQcyrk9CylytMEv3X_YyPAsWTnI84L) where I restarted the R session before each groupComparison run with and without the progress file writing and got the following preliminary results (8 runs):
    - With file progress tracker: 14.67 seconds mean, 15.04 seconds max, 14.14 seconds min
    - Without file progress tracker: 14.48 seconds mean, 15.03 seconds max, 13.85 seconds min

**Conclusion**: File writing impact is minimal.  At most ~5% speed decrease and as low as 1% speed decrease


### Error & Warning Handling

- See this [commit](https://github.com/Vitek-Lab/MSstats/commit/b8b123472cc958880f74d84ca145c80979085293) where I added fake errors within a try statement and suppressed warnings.  Verified parLapply didn't throw errors and the output was the same as without errors/warnings.