@mstaniak 2024-02-22
So using environment() passes everything in the global environment to each thread, right?
Is there a way to pass less? Not sure if it makes much difference tbh.
Otherwise the changes are OK. Please consider the question above, but I will approve the current version