@tonywu1999 2024-02-22
> @tonywu1999 So using environment() passes everything in the global environment to each thread, right? Is there a way to pass less? Not sure if it makes much difference tbh. Otherwise the changes are OK. Please consider the question above, but I will approve the current version. Also, what's the result of Windows tests?

@mstaniak 

Using the environment() function allows us to pass any variable/function from the current environment (not global) to each thread.  We can control which variables we pass - see [lines 460-462](https://github.com/Vitek-Lab/MSstats/pull/110/files#diff-c5d36063f1c31fc4ab10e25cc5155ea5cfc7209b1b4f03d78f3992a6f268a41fR460-R462) where we only pass a predefined set of variables through the `parallel::clusterExport` function.

Still waiting on tests from Devon on the Windows results.