@mstaniak 2024-02-23
@tonywu1999 thank you very much, it's all good. Let's see if it works on Windows. If there's any issues, we will just modify how users can control the parallel backend.