@tonywu1999 2024-02-23
@mstaniak 

Devon tested on his windows laptop, and the amount of time to process increased for a 1GB dataset from 140s to 170s.  This [article](https://www.inwt-statistics.com/blog/code-performance-in-r-parallelization) explains that on Linux/Mac, each thread reads from the same single copy of a dataset, whereas for Windows, each process needs its own copy of the dataset, leading to increased time.

For a sanity check, I tested on the Northeastern HPC, which is run on Linux, and performance improved for a 2GB dataset from 71s to 31s.  For now, I updated the documentation to mention that the parallel processing only works for Mac/Linux.